Conflicts (manually resolved: use pristine file from tarball):
	configure

+++ gw/smscconn_p.h
@@ -1,7 +1,7 @@
 /* ==================================================================== 
  * The Kannel Software License, Version 1.0 
  * 
- * Copyright (c) 2001-2004 Kannel Group  
+ * Copyright (c) 2001-2005 Kannel Group  
  * Copyright (c) 1998-2001 WapIT Ltd.   
  * All rights reserved. 
  * 
@@ -194,7 +194,7 @@ struct smscconn {
 
     int alt_dcs; /* use alternate DCS 0xFX */
 
-    int throughput;     /* message thoughput per sec. to be delivered to SMSC */
+    double throughput;     /* message thoughput per sec. to be delivered to SMSC */
 
     /* Stores rerouting information for this specific smsc-id */
     int reroute;                /* simply turn MO into MT and process internally */
